requirements.txt
asgiref==3.8.1
Django==5.1.1
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pytesseract==0.3.13
sqlparse==0.5.1
tzdata==2024.2
